Implement the main game() function
All the functionality to the game is implemented. The player can choose
either rock, paper, or scissors and battle the computer who will
randomly choose either rock, paper, or scissors. The player is continually
prompted until they complete the 5 rounds. It ends with either the player
winning, losing, or drawing.